requirements.txt
absl-py==0.15.0
amqp==5.1.0
ansicon==1.89.0
arrow==1.2.2
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
blessed==1.19.1
cachetools==5.0.0
celery==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
clang==5.0
cleantext==1.1.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
Django==4.0.4
django-crontab==0.7.1
django-filter==22.1
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
Flask==2.1.2
flatbuffers==1.12
gast==0.4.0
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.1.0
helper==2.5.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.12.1
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.1.0
jupyter-client==7.2.1
jupyter-core==4.9.2
keras==2.8.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kombu==5.2.4
libclang==14.0.1
lxml==4.9.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.15.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
waitress==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
whitenoise==6.2.0
wrapt==1.12.1
zipp==3.8.0
